split syslog go files
